refactor(web): extract useSessionSearch hook
Move the session-search domain out of App: query/filters state, the
local-vs-server result engine, keyboard-selection and result-scroll
effects, and semantic actions (open/submit/close). The hook exposes
refresh() for the live-update subscription; the global keydown handler
stays in App and drives selection via the returned state and setters.
Third step of the App.tsx decomposition.
